+++ dev/io.openliberty.security.jakartasec.4.0.internal_fat/test-applications/ham/src/multiple/ham/inbuilt/MultipleHAMQualifiersApplication.java
@@ -10,7 +10,7 @@
  * Contributors:
  * IBM Corporation - initial API and implementation
  *******************************************************************************/
-package multiple.ham.inbuilt.qualifiers;
+package multiple.ham.inbuilt;
 
 import jakarta.security.enterprise.authentication.mechanism.http.BasicAuthenticationMechanismDefinition;
 import jakarta.security.enterprise.authentication.mechanism.http.CustomFormAuthenticationMechanismDefinition;
